services/sensor-network-management/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
sqlalchemy==2.0.25
asyncpg==0.29.0
redis==5.0.1
influxdb-client==1.39.0
numpy==1.26.3
scipy==1.11.4
scikit-learn==1.4.0
pandas==2.1.4
geopandas==0.14.2
shapely==2.0.2
pyproj==3.6.1
httpx==0.26.0
python-multipart==0.0.6
ortools==9.8.3296